Fix types and cpu platform defines (#7)

+++ public/steam/steamtypes.h
@@ -64,8 +64,8 @@ typedef short int16;
 typedef unsigned short uint16;
 typedef int int32;
 typedef unsigned int uint32;
-typedef int64_t int64;
-typedef uint64_t uint64;
+typedef long long int64;
+typedef unsigned long long uint64;
 
 // [u]int64 are actually defined as 'long long' and gcc 64-bit
 // doesn't automatically consider them the same as 'long int'.
@@ -77,11 +77,11 @@ typedef long int lint64;
 typedef unsigned long int ulint64;
 
 #ifdef X64BITS
-typedef int64_t intp;
-typedef uint64_t uintp;
+typedef long long intp;
+typedef unsigned long long uintp;
 #else
-typedef int32_t intp;
-typedef uint32_t uintp;
+typedef int intp;
+typedef unsigned int uintp;
 #endif
 
 #endif // else _WIN32
